requirements.txt
asgiref==3.5.2
certifi==2022.12.7
distlib==0.3.6
Django==4.1.2
django-debug-toolbar==3.7.0
django-filter==22.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
filelock==3.8.2
mysqlclient==2.1.1
pipenv==2022.12.19
platformdirs==2.6.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.6
virtualenv==20.17.1
virtualenv-clone==0.5.7
